renamed package of puzzleClassifier -> puzzleMaker
